# Performance improvement: Using uncompressed series for temporary series on ApplySerieRangeIntoNewSerie